Test cleanup: Move featuregate-related utils to the config package
This commit moves DisableFeatureGate and EnableFeatureGate (and related utils) from tests/utils.go to the already existing config package.
