+++ tests/BUILD.bazel
@@ -18,10 +18,8 @@ go_library(
         "//tests/console:go_default_library",
         "//tests/exec:go_default_library",
         "//tests/flags:go_default_library",
-        "//tests/framework/checks:go_default_library",
         "//tests/framework/kubevirt:go_default_library",
         "//tests/framework/matcher:go_default_library",
-        "//tests/libkubevirt:go_default_library",
         "//tests/libnode:go_default_library",
         "//tests/libpod:go_default_library",
         "//tests/testsuite:go_default_library",
@@ -191,6 +189,7 @@ go_test(
         "//tests/libinstancetype:go_default_library",
         "//tests/libinstancetype/builder:go_default_library",
         "//tests/libkubevirt:go_default_library",
+        "//tests/libkubevirt/config:go_default_library",
         "//tests/libmigration:go_default_library",
         "//tests/libnet:go_default_library",
         "//tests/libnet/cloudinit:go_default_library",
