+++ tests/launchsecurity/BUILD.bazel
@@ -17,6 +17,7 @@ go_library(
         "//tests/framework/checks:go_default_library",
         "//tests/framework/kubevirt:go_default_library",
         "//tests/framework/matcher:go_default_library",
+        "//tests/libkubevirt/config:go_default_library",
         "//tests/libpod:go_default_library",
         "//tests/libvmifact:go_default_library",
         "//tests/libvmops:go_default_library",
