+++ tests/libkubevirt/config/featuregate.go
@@ -1,0 +1,116 @@
+/*
+ * This file is part of the KubeVirt project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Copyright The KubeVirt Authors.
+ *
+ */
+
+package config
+
+import (
+	"context"
+	"time"
+
+	. "github.com/onsi/gomega"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	v1 "kubevirt.io/api/core/v1"
+	"kubevirt.io/client-go/log"
+
+	"kubevirt.io/kubevirt/tests/flags"
+	"kubevirt.io/kubevirt/tests/framework/checks"
+	"kubevirt.io/kubevirt/tests/framework/kubevirt"
+	"kubevirt.io/kubevirt/tests/libkubevirt"
+	"kubevirt.io/kubevirt/tests/testsuite"
+)
+
+func DisableFeatureGate(feature string) {
+	if !checks.HasFeature(feature) {
+		return
+	}
+	virtClient := kubevirt.Client()
+
+	kv := libkubevirt.GetCurrentKv(virtClient)
+	if kv.Spec.Configuration.DeveloperConfiguration == nil {
+		kv.Spec.Configuration.DeveloperConfiguration = &v1.DeveloperConfiguration{
+			FeatureGates: []string{},
+		}
+	}
+
+	var newArray []string
+	featureGates := kv.Spec.Configuration.DeveloperConfiguration.FeatureGates
+	for _, fg := range featureGates {
+		if fg == feature {
+			continue
+		}
+
+		newArray = append(newArray, fg)
+	}
+
+	kv.Spec.Configuration.DeveloperConfiguration.FeatureGates = newArray
+	if checks.RequireFeatureGateVirtHandlerRestart(feature) {
+		updateKubeVirtConfigValueAndWaitHandlerRedeploymnet(kv.Spec.Configuration)
+		return
+	}
+
+	UpdateKubeVirtConfigValueAndWait(kv.Spec.Configuration)
+}
+
+func EnableFeatureGate(feature string) *v1.KubeVirt {
+	virtClient := kubevirt.Client()
+
+	kv := libkubevirt.GetCurrentKv(virtClient)
+	if checks.HasFeature(feature) {
+		return kv
+	}
+
+	if kv.Spec.Configuration.DeveloperConfiguration == nil {
+		kv.Spec.Configuration.DeveloperConfiguration = &v1.DeveloperConfiguration{
+			FeatureGates: []string{},
+		}
+	}
+
+	kv.Spec.Configuration.DeveloperConfiguration.FeatureGates = append(kv.Spec.Configuration.DeveloperConfiguration.FeatureGates, feature)
+
+	if checks.RequireFeatureGateVirtHandlerRestart(feature) {
+		return updateKubeVirtConfigValueAndWaitHandlerRedeploymnet(kv.Spec.Configuration)
+	}
+
+	return UpdateKubeVirtConfigValueAndWait(kv.Spec.Configuration)
+}
+
+func updateKubeVirtConfigValueAndWaitHandlerRedeploymnet(kvConfig v1.KubeVirtConfiguration) *v1.KubeVirt {
+	virtClient := kubevirt.Client()
+	ds, err := virtClient.AppsV1().DaemonSets(flags.KubeVirtInstallNamespace).Get(context.TODO(), "virt-handler", metav1.GetOptions{})
+	Expect(err).ToNot(HaveOccurred())
+	currentGen := ds.Status.ObservedGeneration
+	kv := testsuite.UpdateKubeVirtConfigValue(kvConfig)
+	Eventually(func() bool {
+		ds, err := virtClient.AppsV1().DaemonSets(flags.KubeVirtInstallNamespace).Get(context.TODO(), "virt-handler", metav1.GetOptions{})
+		Expect(err).ToNot(HaveOccurred())
+		gen := ds.Status.ObservedGeneration
+		if gen > currentGen {
+			return true
+		}
+		return false
+
+	}, 90*time.Second, 1*time.Second).Should(BeTrue())
+
+	waitForConfigToBePropagated(kv.ResourceVersion)
+	log.DefaultLogger().Infof("system is in sync with kubevirt config resource version %s", kv.ResourceVersion)
+
+	return kv
+}
