+++ tests/migration/BUILD.bazel
@@ -46,6 +46,7 @@ go_library(
         "//tests/libdv:go_default_library",
         "//tests/libinfra:go_default_library",
         "//tests/libkubevirt:go_default_library",
+        "//tests/libkubevirt/config:go_default_library",
         "//tests/libmigration:go_default_library",
         "//tests/libnet:go_default_library",
         "//tests/libnet/job:go_default_library",
